`mark_dead` NetworkStream method to support client keep-alives.
When the `with_connector` client constructor is used, and those connections are being
retrieved from a pool, and re-used as keep-alive connections, the stream
needs some way to be notified that it is going to be closed by the
server because "Connection: keep-alive" was not provided in the
response.

Then, the stream (which will be closed by the server and the next
request would fail) can be `drop()`-ped instead of being retained in the
pool.

Conflicts:
	src/header/common/util.rs
	src/header/mod.rs